requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
motor==3.3.2  
tenacity==8.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aws-requests-auth==0.4.3
python-amazon-sp-api==0.19.0
httpx==0.25.2
aiohttp==3.9.1
boto3==1.34.0 
pandas==2.0.3
numpy==1.24.4

redis==5.0.1
aioredis==2.0.1

structlog==23.2.0
prometheus-client==0.19.0

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0